# api/openapi.yaml
openapi: 3.0.3
info:
  title: LabLDAP Management API
  version: 1.0.0
  description: |
    Versioned control-plane REST contract for LabLDAP.

    OpenAPI 3.0.3 is the documented compatible subset of 3.1 (OD-009).
    Payload shapes follow design §3.8 (`x-labldap-contract: proposed`
    stand-in dated 2026-08-13). Architecture-named (binding) routes are
    unmarked. Every other operation is marked `x-labldap-contract: proposed`.

    There is no `PATCH /api/v1/groups/{id}` in v1. Membership mutations
    are the group update path.

    Examples use lab placeholders only. They are not usable credentials.
  license:
    name: UNLICENSED
    url: https://github.com/hilather/go-lab-ldap-mcp
...
servers:
  - url: /
    description: Same-origin control listener
...
paths:
...
  /api/v1/users:
    get:
      operationId: listUsers
      tags: [Users]
      summary: List users
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
        - name: q
          in: query
          schema:
            type: string
          description: Optional substring on id or uid
      responses:
        "200":
          description: User page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      tags: [Users]
      summary: Create a user
      description: Binding `POST /api/v1/users`. Returns 201, Location, and ETag.
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSpec"
            example:
              id: alice
              uid: alice
              enabled: true
              password: lab-example-password
              attributes: {}
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
                example: /api/v1/users/alice
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      operationId: getUser
      tags: [Users]
      summary: Get a user
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      responses:
        "200":
          description: User
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateUser
      tags: [Users]
      summary: Patch a user (not password)
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Updated
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Delete a user
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
...
  /api/v1/groups:
    get:
      operationId: listGroups
      tags: [Groups]
      summary: List groups
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Group page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createGroup
      tags: [Groups]
      summary: Create a group
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupSpec"
            example:
              id: staff
              members:
                - kind: user
                  id: alice
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
                example: /api/v1/groups/staff
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Static lab token from a secret file (OD-013). Compared in constant
        time. Only the non-secret token ID is exposed after authentication.
        Scopes: directory:read, directory:write, directory:password,
        lab:reset, lab:export, schema:read, audit:read.
    cookieAuth:
      type: apiKey
      in: cookie
      name: labldap_session
      description: |
        Opaque in-memory session cookie. HttpOnly, SameSite=Strict,
        Secure when TLS, Path=/. Cookie-authenticated mutations require
        header X-CSRF-Token and a valid Origin.
  parameters:
    UserID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: Quoted revision hex
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      schema:
        type: string
      description: Required for cookie-authenticated unsafe methods
  responses:
    BadRequest:
      description: Invalid JSON or field error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, malformed, or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://labldap.dev/problems/auth
            title: authentication required
            status: 401
            extensions:
              requestId: 00000000000000000000000000000000
    Forbidden:
      description: Missing scope, CSRF, or Origin
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: Unknown resource
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Duplicate or in-progress operation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PreconditionFailed:
      description: If-Match revision mismatch
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    AttrKV:
      type: object
      additionalProperties: false
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
    User:
      type: object
      additionalProperties: false
      required: [id, uid, dn, enabled, objectClasses, attributes, revision]
      properties:
        id:
          type: string
        uid:
          type: string
        dn:
          type: string
        enabled:
          type: boolean
        objectClasses:
          type: array
          items:
            type: string
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/AttrKV"
        groups:
          type: array
          items:
            type: string
        revision:
          type: string
    UserSpec:
      type: object
      additionalProperties: false
      required: [id, password]
      properties:
        id:
          type: string
        uid:
          type: string
        enabled:
          type: boolean
        password:
          type: string
          format: password
        attributes:
          type: object
          additionalProperties:
            type: string
    UserPatch:
      type: object
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        attributes:
          type: object
          additionalProperties:
            type: string
    UserPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        nextCursor:
          type: string
    MemberRef:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind:
          type: string
          enum: [user, group]
        id:
          type: string
        dn:
          type: string
...
    Group:
      type: object
      additionalProperties: false
      required: [id, dn, members, revision]
      properties:
        id:
          type: string
        dn:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberRef"
        revision:
          type: string
    GroupSpec:
      type: object
      additionalProperties: false
      required: [id, members]
      properties:
        id:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberRef"
    GroupPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        nextCursor:
          type: string
...
    FieldError:
      type: object
      additionalProperties: false
      required: [path, code, message]
      properties:
        path:
          type: string
        code:
          type: string
        message:
          type: string
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://labldap.dev/problems/auth
        title:
          type: string
        status:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        extensions:
          type: object
          additionalProperties: false
          properties:
            requestId:
              type: string
security:
  - bearerAuth: []
  - cookieAuth: []